Spotify User Analysis
This report explores a single Spotify user’s data to extract insights and define a user profile. A recommendation model is designed giving consideration to multiple algorithms and hypothesis’. The end result is a profile with which one can query song data to feed to the recommendation model to predict which songs the user will like. Model performance statistics are laded out on the last page.